getBodyStringOfFunction
