+++ compiler/examples/template-api-examples/nba_game_table.js
@@ -1,12 +1,12 @@
 // libraries
-const Project = require("../../../src/index").Project;
-const Canvas = require("../../../src/Canvas").Canvas;
-const Jump = require("../../../src/Jump").Jump;
-const Layer = require("../../../src/Layer").Layer;
-const View = require("../../../src/View").View;
-const Table = require("../../../src/template-api/Table").Table;
+const Project = require("../../src/index").Project;
+const Canvas = require("../../src/Canvas").Canvas;
+const Jump = require("../../src/Jump").Jump;
+const Layer = require("../../src/Layer").Layer;
+const View = require("../../src/View").View;
+const Table = require("../../src/template-api/Table").Table;
 
-var p = new Project("epl", "../../../../config.txt");
+var p = new Project("epl", "../../../config.txt");
 
 var db = "nba";
 var fields = [
